requirements-base.txt
django==3.1.6
djangorestframework==3.11.0
django-profanity-filter==0.2.1
black==21.5b1